Fix updating model access with new models
