+++ src/TrainingActivity2.py
@@ -1,103 +1,18 @@
-import argparse
-from collections import deque
-from datetime import datetime, timedelta
-from time import sleep
-
-from src.hyperparameters import Hyperparameters
-from src.novelty.NoveltyName import NoveltyName, noveltyList
-from src.novelty.NoveltyDirector import NoveltyDirector
-from src.exceptions.NoModelException import NoModelException
-from src.training.ModelAccess import ModelAccess
-from src.training.ModelTraining import trainNewModel, train_best, train_last, train_model
-from src.training.TuningParameters import scale_learning_rate, scale_batch_size
-
-
-class TrainingActivity:
-    def __init__(self,
-                 novelty_name: NoveltyName = NoveltyName.ORIGINAL,
-                 render=None,
-                 continuous=True,
-                 train_hour=8):
-        self.novelty_name = novelty_name
-        self.render = render
-        self.continuous = continuous
-        self.train_hour = train_hour
-        self.model_access = ModelAccess(novelty_name)
-
-    def __del__(self):
-        try:
-            del self.model_access
-        except AttributeError:
-            print("model_access object not created.")
-
-
-    def train(self):
-        # setting up environment
-        env = NoveltyDirector(self.novelty_name).build_env(self.render, self.continuous)
-        model_access = ModelAccess(self.novelty_name)
-
-        # dynamic updatable params
-        params: Hyperparameters = None
-        recent_success = []
-        recent_files = []
-        queue_size = 10
-        prev_success = False
-
-        start_time = datetime.now()
-        end_time = start_time + timedelta(hours=self.train_hour)
-        while datetime.now() < end_time:
-            try:
-                if prev_success:
-                    name, model = model_access.get_latest()
-                else:
-                    name, model = model_access.get_best()
-                result = train_model(env=env,
-                                     prev_model=model,
-                                     prev_filename=name,
-                                     params=params,
-                                     model_novelty=self.novelty_name)
-                model_access.add_model(name, model)
-
-                # Update records
-                params = result.params
-                prev_success = result.success
-                recent_success.append(result.success)
-                recent_files.append(result.filename)
-                if len(recent_success) > queue_size:
-                    recent_success.pop(0)
-                    recent_files.pop(0)
-
-                # Adjust hyper parameters
-                # if the same file is training all the time
-                count = [recent_files.count(s) for s in set(recent_files)]
-                if len(count) == 1:
-                    # explore
-                    scale_learning_rate(params, 1.5)
-                    scale_batch_size(params, 0.25)
-                elif any([x > queue_size / 2 for x in count]) or result.success:
-                    # exploit
-                    scale_learning_rate(params, 0.95)
-                    scale_batch_size(params, 2)
-                else:
-                    # explore
-                    scale_learning_rate(params, 1.5)
-                    scale_batch_size(params, 0.25)
-
-                sleep(20)
-            except NoModelException:
-                trainNewModel(env, self.novelty_name)
+from src.novelty.NoveltyName import NoveltyName
+from src.training.Training import Training
 
 
 if __name__ == "__main__":
     try:
-        activity = TrainingActivity(
+        activity = Training(
             novelty_name=NoveltyName.SENSOR,
             render=None,
             continuous=True,
-            train_hour=4)
+            train_hour=1)
+        activity.train()
     except KeyboardInterrupt:
         print("Training activity interrupted.")
         del activity
         print("Destructor finished.")
-
-
+    finally:
+        del activity
